requirements.txt
asgiref==3.5.2
Django==4.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mysqlclient==2.1.1
Pillow==9.3.0
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.7

keras~=2.10.0
numpy~=1.24.2
tensorflow~=2.10.1
scikit-image~=0.19.3